Jwt token based auth for engine api
