+++ packages/client/bin/startRpc.ts
@@ -1,0 +1,193 @@
+import { Server as RPCServer } from 'jayson/promise'
+import { json as jsonParser } from 'body-parser'
+import Express, { Application, Request, Response, NextFunction } from 'express'
+import jwt from 'express-jwt'
+
+import { RPCManager } from '../lib/rpc'
+import EthereumClient from '../lib/client'
+import { inspectParams } from '../lib/util'
+import * as modules from '../lib/rpc/modules'
+
+type RPCArgs = {
+  rpc: boolean
+  rpcaddr: string
+  rpcport: number
+  ws: boolean
+  wsPort: number
+  wsAddr: string
+  rpcEngine: boolean
+  rpcEngineAddr: string
+  rpcEnginePort: number
+  rpcDebug: boolean
+  helprpc: boolean
+  'jwt-secret': string
+}
+
+function createRPCServerListener({
+  rpcport,
+  server,
+  withEngineMiddleware,
+}: {
+  rpcport: number
+  server: RPCServer
+  withEngineMiddleware?: { jwtEngineSecret: string }
+}): Application {
+  const app = Express()
+  app.use(jsonParser())
+  if (withEngineMiddleware) {
+    app.use(
+      jwt({ secret: withEngineMiddleware.jwtEngineSecret, algorithms: ['HS256'] }).unless(function (
+        req: Request
+      ) {
+        const { method } = req.body
+        return !method.includes('engine_')
+      })
+    )
+    app.use(function (req: Request, res: Response, next: NextFunction) {
+      const { user } = req as unknown as { user: { iat: number } | undefined }
+      if (user && Math.abs(new Date().getTime() - user.iat * 1000 ?? 0) > 5000) {
+        console.log('expired', user)
+        return res.sendStatus(401)
+      }
+      return next()
+    })
+  }
+  app.use(server.middleware())
+  app.listen(rpcport)
+  return app
+}
+
+/**
+ * Starts and returns enabled RPCServers
+ */
+export function startRPCServers(client: EthereumClient, args: RPCArgs) {
+  const config = client.config
+  const onRequest = (request: any) => {
+    let msg = ''
+    if (args.rpcDebug) {
+      msg += `${request.method} called with params:\n${inspectParams(request.params)}`
+    } else {
+      msg += `${request.method} called with params: ${inspectParams(request.params, 125)}`
+    }
+    config.logger.debug(msg)
+  }
+
+  const handleResponse = (request: any, response: any, batchAddOn = '') => {
+    let msg = ''
+    if (args.rpcDebug) {
+      msg = `${request.method}${batchAddOn} responded with:\n${inspectParams(response)}`
+    } else {
+      msg = `${request.method}${batchAddOn} responded with: `
+      if (response.result) {
+        msg += inspectParams(response, 125)
+      }
+      if (response.error) {
+        msg += `error: ${response.error.message}`
+      }
+    }
+    config.logger.debug(msg)
+  }
+
+  const onBatchResponse = (request: any, response: any) => {
+    // Batch request
+    if (request.length !== undefined) {
+      if (response.length === undefined || response.length !== request.length) {
+        config.logger.debug('Invalid batch request received.')
+        return
+      }
+      for (let i = 0; i < request.length; i++) {
+        handleResponse(request[i], response[i], ' (batch request)')
+      }
+    } else {
+      handleResponse(request, response)
+    }
+  }
+
+  const servers: RPCServer[] = []
+  const {
+    rpc,
+    rpcaddr,
+    rpcport,
+    ws,
+    wsPort,
+    wsAddr,
+    rpcEngine,
+    rpcEngineAddr,
+    rpcEnginePort,
+    'jwt-secret': jwtEngineSecret,
+  } = args
+  const manager = new RPCManager(client, config)
+
+  if (rpc || ws) {
+    const withEngineMethods = rpcEngine && rpcEnginePort === rpcport && rpcEngineAddr === rpcaddr
+    const methods = withEngineMethods
+      ? { ...manager.getMethods(), ...manager.getMethods(true) }
+      : { ...manager.getMethods() }
+    const server = new RPCServer(methods)
+    server.on('request', onRequest)
+    server.on('response', onBatchResponse)
+    const namespaces = [...new Set(Object.keys(methods).map((m) => m.split('_')[0]))].join(',')
+    if (rpc) {
+      createRPCServerListener({
+        rpcport,
+        server,
+        withEngineMiddleware: withEngineMethods ? { jwtEngineSecret } : undefined,
+      })
+      config.logger.info(
+        `Started JSON RPC Server address=http://${rpcaddr}:${rpcport} namespaces=${namespaces}`
+      )
+    }
+    if (ws) {
+      const opts: any = { port: wsPort }
+      if (rpcaddr === wsAddr && rpcport === wsPort) {
+        // If http and ws are listening on the same port,
+        // pass in the existing server to prevent a listening error
+        delete opts.port
+        opts.server = server
+      }
+      server.websocket(opts)
+      config.logger.info(
+        `Started JSON RPC Server address=ws://${wsAddr}:${wsPort} namespaces=${namespaces}`
+      )
+    }
+    servers.push(server)
+  }
+
+  if (rpcEngine) {
+    if (rpc && rpcport === rpcEnginePort && rpcaddr === rpcEngineAddr) {
+      return servers
+    }
+    const server = new RPCServer(manager.getMethods(true))
+    server.on('request', onRequest)
+    server.on('response', onBatchResponse)
+
+    createRPCServerListener({ rpcport, server, withEngineMiddleware: { jwtEngineSecret } })
+    config.logger.info(
+      `Started JSON RPC server address=http://${rpcEngineAddr}:${rpcEnginePort} namespaces=engine`
+    )
+
+    servers.push(server)
+  }
+
+  return servers
+}
+
+/**
+ * Output RPC help and exit
+ */
+export function helprpc() {
+  console.log('-'.repeat(27))
+  console.log('JSON-RPC: Supported Methods')
+  console.log('-'.repeat(27))
+  console.log()
+  for (const modName of modules.list) {
+    console.log(`${modName}:`)
+    const methods = RPCManager.getMethodNames((modules as any)[modName])
+    for (const methodName of methods) {
+      console.log(`-> ${modName.toLowerCase()}_${methodName}`)
+    }
+    console.log()
+  }
+  console.log()
+  process.exit()
+}
