+++ packages/client/package.json
@@ -57,9 +57,12 @@
     "@ethereumjs/ethash": "^1.1.0",
     "@ethereumjs/tx": "^3.5.0",
     "@ethereumjs/vm": "^5.7.1",
+    "body-parser": "^1.19.2",
     "chalk": "^4.1.2",
     "debug": "^4.3.3",
     "ethereumjs-util": "^7.1.4",
+    "express": "^4.17.3",
+    "express-jwt": "^6.1.0",
     "fs-extra": "^10.0.0",
     "it-pipe": "^1.1.0",
     "it-pushable": "^1.4.2",
@@ -83,6 +86,8 @@
   },
   "devDependencies": {
     "@babel/plugin-transform-spread": "^7.10.1",
+    "@types/express": "^4.17.13",
+    "@types/express-jwt": "^6.0.4",
     "@types/fs-extra": "^8.1.0",
     "@types/levelup": "^4.3.0",
     "@types/node": "^16.11.7",
